Fix compiler warnings on macOS in CardDetectionThread().
ccid_usb.c:2385:21: warning: cast to smaller integer type 'int' from
'void *' [-Wvoid-pointer-to-int-cast]
        int reader_index = (int) pParam;
                           ^~~~~~~~~~~~

+++ src/ccid_usb.c
@@ -2387,7 +2387,7 @@ static struct usbDevice_MultiSlot_Extension *Multi_CreateNextSlot(int physical_r
 // Card detection thread
 static void *CardDetectionThread(void *pParam)
 {
-	int reader_index = (int) pParam;
+	int reader_index = (int) (intptr_t) pParam;
 	int rv = 0;
 	int status = 0;
 	int actual_length = 0;
